[FEAT] Scheduler / snapshot 저장, Summary 스케쥴러  (#123)
* feat: scheduler, 단기예보 저장 시간 조정 (26H > 36H)

* fix: 단기예보 6시간에 한번씩 호출

* feat: snapshot scheduler 구현완료, (snapshot의 sky, pty를 int 로 변경)

* feat: snapshot scheduler에 체감온도 추가하는 공식 추가

* chore: test 변경, jpa 설정 변경

* chore: 불필요 속성 제거

* feat: daily Summary Scheduler 구현완료

* test: dailySummary 테스트

+++ src/main/java/com/weady/weady/domain/tags/dto/TemperatureTagResponseDto.java
@@ -5,6 +5,6 @@ import lombok.Builder;
 @Builder
 public record TemperatureTagResponseDto(Long id,
                                         String name,
-                                        Integer minTemperature,
-                                        Integer maxTemperature) {
+                                        Float minTemperature,
+                                        Float maxTemperature) {
 }
